+++ cli-tool/app/commands/export.rb
@@ -1,20 +1,14 @@
 module Taylor
   module Command
     class Export
-      def self.call(parser)
-        self.new(parser)
+      def self.call(argv, options)
+        self.new(argv, options)
       end
 
       attr_accessor :options
 
-      def initialize(parser)
-        @options = parser.opts.slice(
-          :help, :dry_run, :name, :input, :export_directory, :load_paths,
-          :copy_paths
-        )
-
-        @options[:load_paths] = @options[:load_paths].split(',')
-        @options[:copy_paths] = @options[:copy_paths].split(',')
+      def initialize(argv, options)
+        setup_options(argv, options)
 
         if @options[:help]
           display_help
@@ -49,6 +43,23 @@ module Taylor
       end
 
       private
+      def setup_options(argv, options)
+        parser = OptParser.new do |opts|
+          opts.on(:help,             :bool,   false)
+          opts.on(:dry_run,          :bool,   false)
+          opts.on(:name,             :string, options.fetch(:name,             'taylor_game'))
+          opts.on(:input,            :string, options.fetch(:input,            'game.rb'))
+          opts.on(:export_directory, :string, options.fetch(:export_directory, './exports'))
+          opts.on(:load_paths,       :string, options.fetch(:load_paths,       './,./vendor'))
+          opts.on(:copy_paths,       :string, options.fetch(:copy_paths,       './assets'))
+        end
+        parser.parse(argv, true)
+
+        @options = parser.opts
+        @options[:load_paths] = @options[:load_paths].split(',')
+        @options[:copy_paths] = @options[:copy_paths].split(',')
+      end
+
       def check_in_taylor_project!
         unless File.exists?(File.join(Dir.pwd, 'taylor-config.json'))
           raise 'This command must be run from a Taylor project'