+++ cli-tool/app/commands/run.rb
@@ -1,15 +1,17 @@
 module Taylor
   module Command
     class Run
-      def self.call(command, parser)
-        self.new(command, parser)
+      def self.call(command, argv, options)
+        self.new(command, argv, options)
       end
 
-      def initialize(command, parser)
-        @command = command
-        @from_config = parser.opts[:input]
+      attr_accessor :options
+      def initialize(command, argv, options)
+        setup_options(argv, options)
 
-        if parser.opts[:help]
+        @command = command || '.'
+
+        if @options[:help]
           display_help
         else
           call
@@ -27,12 +29,14 @@ module Taylor
           Taylor #{TAYLOR_VERSION}
 
           Usage:
-            taylor\t\t# If the current folder is a taylor game, launch it
-            taylor ./game.rb\t# Launch ./game.rb
+            taylor\t\t\t# If the current folder is a taylor game, launch it
+            taylor ./game.rb\t\t# Launch ./game.rb
+            taylor --input ./game.rb\t# Launch ./game.rb
             taylor [action] [options]
 
           Options:
             --help\tShow this message
+            --input input\t\t\tWhat is the name of the entrypoint file (defaults to game.rb)
 
           Actions:
             new\tCreate a new Taylor game
@@ -40,6 +44,20 @@ module Taylor
       end
 
       private
+      def setup_options(argv, options)
+        parser = OptParser.new do |opts|
+          opts.on(:help,             :bool,   false)
+          opts.on(:input,            :string, options.fetch(:input,            'game.rb'))
+        end
+        parser.parse(argv, true)
+
+        @options = parser.opts
+      end
+
+      def from_config
+        @options[:input]
+      end
+
       def unload_taylor_cli
         Taylor.send(:remove_const, :Command)
       end
@@ -49,11 +67,11 @@ module Taylor
           ARGV.shift
           require @command
 
-        elsif File.exists?(@from_config) && File.file?(@from_config)
-          require @from_config
+        elsif File.exists?(from_config) && File.file?(from_config)
+          require from_config
 
         else
-          raise "Did not know how to handle #{command}"
+          raise "Did not know how to handle #{@command}"
         end
       end
     end